requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
gunicorn>=22.0.0,<24.0.0

langgraph>=0.6.0,<0.7.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langsmith>=0.3.0,<0.4.0

pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0

azure-identity>=1.14.0

PyPDF2>=3.0.0,<4.0.0
pandas>=2.0.0,<3.0.0

duckduckgo-search>=6.0.0,<8.0.0
